+++ stories_manager/src/main/java/com/sweng_stories/stories_manager/domain/IndovinelloFactory.java
@@ -3,7 +3,7 @@ package com.sweng_stories.stories_manager.domain;
 
 public class IndovinelloFactory {
 
-    public static Indovinello createIndovinello(String tipo, Long id, String descrizione, String domanda, Object rispostaCorretta, Long scenarioId) {
+    public static Indovinello createIndovinello(String tipo, int id, String descrizione, String domanda, Object rispostaCorretta, int scenarioId) {
         if ("testuale".equals(tipo)) {
             return new IndovinelloTestuale(id, descrizione, domanda, (String) rispostaCorretta, scenarioId);
         } else if ("numerico".equals(tipo)) {
